Add DACs to support PID/VID verfification (#14845)
* Add new certs for development.

We are moving to have the certificate verification check the VID and PID
between the basic cluster and the DAC/CD. Right now, none of the examples
pass prorperly because the VIDs and PIDs do not match. In order to
facilitate development while platforms are developing their own
DeviceAttestationCredentialsProvider, we have provided a new set of
development certs that can be used for development only.

This new scheme is backed by the test PAA in attestation/test/. This was
done to reduce the number of changes required to the controllers, which
already contain this PAA in their trusted certs.

The PAI has been changed to omit the PID. This means the we can use a
common PAI cert for all products. The vendor ID for the PAI is 0xFFF1,
which a known test vendor for Matter.

The DACs below are signed by the new PAI and include certs and keys
for PIDs 0x8000-0x801F.

* Use new certificates in attestation.

Test: Can commission linux lighting app using pid 0x8000

* Add new certificate declaration

This new CD will veryify against all products with VID 0xFFF1
and PIDs in the range of 0x8000-0x8063.

Test: Verified on linux lighitng app by forcing app and controller
      to use pid 0x8001

* Update example PIDs.

Please see documentation in docs/examples.

* Add explicit warning for PID/VID mismatch.

* Revert "Revert "Check VID/PID when doing device attestation (#14551)" (#14795)"

This reverts commit ad28f32b9929f564ee3233699c13dd72bfe1d924.

* Fix tests

Use hard coded cert values for tests rather than going through the
example creds provider.

* Update src/controller/AutoCommissioner.cpp

Co-authored-by: Tennessee Carmel-Veilleux <tennessee.carmelveilleux@gmail.com>

* Update src/controller/CHIPDeviceController.h

Co-authored-by: Tennessee Carmel-Veilleux <tennessee.carmelveilleux@gmail.com>

* Restyled by clang-format

* Fix two spelling errors

* Update src/credentials/tests/TestDeviceAttestationCredentials.cpp

Co-authored-by: Evgeny Margolis <emargolis@google.com>

* Use define for CD size.

* YOU get a bracket, YOU get a bracket...

everybody gets a bracket!

* Spelling error.

* Fix P6

Sneaky P6...got by me.

* Put function names in single quotes

I think this will appease the spell checker.

* Fine, spell checker. Here you go.

Co-authored-by: Tennessee Carmel-Veilleux <tennessee.carmelveilleux@gmail.com>
Co-authored-by: Restyled.io <commits@restyled.io>
Co-authored-by: Evgeny Margolis <emargolis@google.com>