Fix ACL Delete Logic
Problem:

When deleting existing ACLs as part of processing a Write Request
containing a 'replace all' list operation, the existing logic was
incorrectly passing a pointer to the last entry that we had processed in
the new provided list when calling LogEvent.

Not only is this logging the wrong entry beyond the first iteration of
the loop, it also does not work if the new list being provided to the caller
is actually empty.

This caused it to send an error back to the client.

Fix:

Correctly read out the i-th entry from the existing ACL list and log
that before deleting that entry.

Testing:

Validated by using reproducing the original error in the REPL, and
validating that it succeeds after.