Install ACL for group messaging test
Group messaging test will of course fail without an installed ACL
that permits the tested IM operations via group auth mode.

Install an ACL during the test so it can succeed.

Also, complete the work to transcode Group and PASE subjects from
the cluster level (not encoded in a NodeId) to the system level
(encoded in a NodeId).

Part of #14985