Don't encode fabric index for fabric-scoped structs when it's set to 0.
For reads, this situation should never arise: the fabric-scoped struct
should always have a valid fabric index.

For writes, this allows a consumer to just not set a fabric index
(which would get ignored anyway), and since the default value is 0
that will lead to no fabric index being sent.

The other option would be to explicitly make the fabric-index
possibly-missing (using Optional or some subclass of Optional), but
since we use the same structs for both encode and decode in many cases
that would force the decoding side to worry about whether the value is
present or not (which it always would be after decoding).