Separate DataModel APIs for encoding fabric-sensitive structs
This builds upon bzbarsky's original PR #15268 based upon discussions
with him to create separate EncodeForWrite/EncodeForRead APIs that only
apply to fabric-scoped structs.

This is then used (for now) to omit encoding the fabric index on writes.

Tests: Ran chip-tool to write an ACL entry and confirmed that that the
fabric index field is omitted on writes.