Add PermissiveAccessControlDelegate for tests
This new delegate does not error and overrides the access control
check to always allow. It is installed in the unit test context
so that unit tests (e.g. of interaction model, which has access
control baked in) can test units other than access control
without failing due to access control, and without running with
real access control and having to install real ACLs (that's for
integration tests to do).