Enforce access control (#14921)
* Enforce access control

Issue #14454

* Forgot to save this file

* Zap regen

* Small logging changes in access control

May not keep this for release but need it for the next few weeks.

* Add another log

* Add temp logs to debug nRF test failures on Zephyr

* Temporarily change ChipLogDetail to ChipLogError

* Change access control logs from detail to progress

* Remove temporary logs

* Add PermissiveAccessControlDelegate for tests

This new delegate does not error and overrides the access control
check to always allow. It is installed in the unit test context
so that unit tests (e.g. of interaction model, which has access
control baked in) can test units other than access control
without failing due to access control, and without running with
real access control and having to install real ACLs (that's for
integration tests to do).

* Fix logging

I meant to do this in the last commit.

* Add test helper dependency on access control