Strongly Typed Bitmaps
This adds support for strongly typed BITMAP* fields in the XML that in
turn result in the use of the BitFlags<T> type in the generated C++
code.

To correctly work this type for attributes, a specialization of
`NumericAttributeTraits` has been added for `BitFlags<T>` to correctly
handle the underlying storage type as well the special rule for
nullability.

Tests:
- Ensured the YAML tests for bitmaps still works.