Revert "Revert "CI split between master (full) and PR (smoke) for ESP32 (#15605)" (#15773)"
This reverts commit 66aef3e0b939b429aa097d25a605dcc6087b582c.

Adding back the split to remove queueing time. Please do not revert and
rather work to adding appropriate splits or logic as done in #15791
or add additional builds as needed.

The split decreases PR queue time (not as affected) and has a global
timeout which seems easier to track than the timeout we are experiencing
in locations like:

https://github.com/project-chip/connectedhomeip/runs/5422883482?check_suite_focus=true