Add timeout argument to test command
The test cluster command in particular is very long and the time
required to run it varies greatly depending on the network setup.
On the M5, we are timing out with the already increased default
value, whereas with a linux device this entire test completes in
less than 10s. Rather than forcing every platform to use the same
default, make the default settable on the command line.

Test: added a log to ensure that the returned GetWaitDuration returns
      the value from the command line. It does, log is removed.