Cleanup dangling CommissioneedeviceProxy object on CASE session establishment failure
This cleans up the dangling CommissioneeDeviceProxy object that is
usually cleaned up in case of successful commissioning in
OnDeviceConnectedFn (i.e successful CASE session establishment upon
completion of commissioning).

Consequently, failure to establish CASE should do the same *if and only
if* the device that we failed to establish CASE with matches the device
being commissioned.

Tests:
- Ensured we hit a VerifyOrDie error due to an open EC if CASE failed to
  establish, and then ensured that didn't manifest again with these
  changes.