Make the test timeout always command-line configurable for YAML tests.
A test's timeout was only configurable if "timeout" was set in the
YAML.  Make sure the codegen acts as if that is always set.

Also modifies zap_regen_all.py to clearly list all the things it's
planning to run to generate things up front, and adds chip-tool-darwin
to the set of things it generates.

Also adds a --dry-run option to zap_regen_all.py to just list the
targets and exit.