[ota] Remove SetOTAImageProcessorParams
The function was initially used to set name of image file
to be downloaded, but most platforms don't use that and
some platforms use the file name as the output file name,
that is, the file name where the downloaded image should
be saved.

Remove the method from the common interface and add
SetOTAImageFile setters in platforms that need to have
the output file name set.