Strongly Typed Bitmaps
Switching some of the attributes in the test cluster of type BITMAP* to
actually using the enum-like definition of the bitmap resulted in all
kinds of compiler errors.

This adds support for strongly typed BITMAP* fields in the XML that in
turn result in the use of the BitFlags type in the generated C++ code.

To correctly work this type for attributes, a specialization of
NumericAttributeTraits has been added for BitFlags<T> to correctly
handle the underlying storage type as well the special rule for
nullability.

Some of other fix-ups to the chip-tool helpers were needed as well.

Tests:
Ensured the YAML tests for bitmaps still works.