TLV to JSON converter (#16086)
This adds a TLV to JSON converter to provide a means to deal with TLV in
a more human-readable format. This does not rely on any schema
information and instead, just dumps out raw TLV present in cluster
payloads into a JSON format.

This relies on the jsoncpp third party library.

Testing
Added a unit-test.