Allow re-establishing pase sessions
The spec indicates that the commissioner should re-establish the
pase session if commissioning fails in most cases. This is not
possible currently because the PASE connection handler is removed
when the initial connection is established. This changes the code
to continue accepting pase connections even after one has been
established.

A second commissioner will be able to establish a PASE connection,
but will not be able to arm the failsafe to start commissioning.
The first commissioner can re-establish and will be able to re-arm
the failsafe because the fabric will match.

The StopAdvertising call still happens after commissinoing complete,
which is the real point when the commissioning window closes.