Speedup REPL script
The 'run_python_test' script was taking way too long to startup on my
machine (upwards of 7-8s).

That was because it was using the click package (which takes about 3-4
to start the application) as well as an extensive path search that was
being done to find the absolute path to the test script.

Fixes:

The former is a known issue with click, so shifted it back to well-test
argparse.

The latter was fixed by just deriving a relative path from where the
current script is running from.

Tests:
- Test now starts immediately, saving 7-8s.