[dns-sd] Extended discovery improvements (#16290)
* [dns-sd] Change extended discovery condition

According to the spec, each time a commissionable node
service is advertised and the device is not in the
commissioning mode, the advertising should be treated as
Extended Discovery. The current code, however, makes
a distinction whether the device is on any fabric or not.

* [dns-sd] Set default extended discovery timeout to 15 minutes

The spec recommends that the extended discovery by default
time out after a period recommended in the "Announcement
duration" section. Change the default from "no timeout" to
15 minutes.

Use DefaultStorageKeyAllocator when persisting the timeout.

Make extended discovery logs more meaningful for a
non-implementer and remove some irrelevant messages.

* Fix Cirque tests

Discriminator must be now explicitly specified in Linux apps.
Previously the Cirque tests were passing only by chance.

* Remove CHIP_DEVICE_CONFIG_ENABLE_COMMISSIONABLE_DISCOVERY

Most platforms have it enabled, by default, and those that
don't, have extended discovery disabled, too, which covers
the case when the commissioning window is closed. On the
other hand, when the commissioning window is open, the
device should always try to advertise over DNS-SD.

* Fix build

* Follow KVS key convention