Fix event handling in commissioning window manager.
The current failsafe expiry doesn't actually re-start the advertising
because the pase connection stops the commissioning window timer.

Instead, stop the commissining window timer when we get a commissioning
complete.

Don't use the OnSessionEstablishmentError function for failsafe
timer expiry because it sends a callback about pairing that does not
apply here.