Added parameter to Objective-C CHIPDeviceController to accept PAA certificates from Objective-C client
Fixes #14126

Change overview

Added a new parameter to CHIPDeviceController that is an array of NSData PAA certificates.
Created an Objective-C bridge for the AttestationTrustStore class. The PAA certificates provided
by the client are passed to the default device attestation verifier. If the client passes a nil
value for the parameter then the testing root store is used.

Testing

Used an Objective-C client to pass valid certificates to ensure device attestation procedure
succeeded with an m5stack device. Passed in nil to ensure the device attestation procedure used
the test PAA certificates.