Add the option to endpoint_config to abort if storage option for attribute is unknown.
This was previous solved in a backwards incompatible way in the zap itself, in process
breaking the previous uses of endpoint_config in ZigbeePro stacks.
With this, the default value is to let unknown options be allowed, but
with adding the allowUnknownStorageOption="false" to the helper, you turn on the strictness.