Ensure we deallocate CASEClient if EstablishSession fails synchronously.
We could end up in a state where we allocated mCASEClient in
OperationalDeviceProxy::EstablishConnection and then EstablishSession
on the CASEClient would fail and we would be left with a stale
pointer.  And then later nothing would deallocate the CASEClient and
if we had another connection attempt we'd just allocate a new one and
leak the old one.

The fix is two-fold:

1. Make sure to call CleanupCASEClient (renamed from CloseCASESession)
   if EstablishSession fails.

2. Make it harder to forget to call CleanupCASEClient in various
   places by just doing it when we transition out of the Connecting
   state.