Stop getting reachable state from configuration manager.
This is dynamic state; reading it once at cluster startup time makes no sense.

In practice, any implementor that wants to implement behavior other
than "return true" here needs to use an external attribute or
AttributeAccessInterface and hook this up to actual reachability data.

The current default value for this in all our endpoint configs is
"true", as expected.