Quick-fix running out of CASE sessions
- From #17422 onwards, "previous" CASE sessions were not expired
  in terms of the session cache if sessions were constantly
  re-established.
- A root cause solution will be to solve #17568, but in the meantime,
  cert testing started failing after N (where N is session manager
  pool size) runs.

This PR cleans-up the previous sessions from same node (duplicate sessions
that should not happen). This is a bit unclean but will do to unblock
testing since it was good enough before, and the better method is not
yet implemented.

This PR also adds logging when this happens.

Fixes #17698

Testing done:
- Unit tests pass
- Cert tests pass
- Manual run of N > 16
  `for i in {1..18}; do ./out/debug/standalone/chip-tool basic read vendor-name 115566 0 ; done`
  does not fail as before.