Simplify implementation of emberAfContainsAttribute.
The current implementation looks up the attribute metadata, which
involves figuring out where in the global attribute storage the
attribute might live, which depends on the state of other endpoints.

Instead, use emberAfGetServerAttributeIndexByAttributeId, which only
depends on the state of the one endpoint in question.

This allows removing a workaround for the "calling these functions
when endpoints are not initialized yet" bits that was added in #17719.