Fix "ethernet" pairing for chip-tool-darwin.
The Matter framework pairDevice API will actually go ahead and
commission the device if it's provided with the IP/port (as opposed to
the versions that just take an onboarding payload, which will set up
PASE but not commission).  The pairing delegate in chip-tool-darwin
was _also_ trying to commission, and then we failed out because we had
two commissioning attempts at the same time.

The fix is to not try to commission in the delegate if we are doing
"ethernet" pairing.