Fix "ethernet" pairing for chip-tool-darwin.
Makes the various pairDevice APIs on CHIPDeviceController have
identical behavior in terms of just establishing a PASE session and
waiting for the API consumer to call commissionDevice as desired.

Without this change, both the API implementation and
chip-tool-darwin's pairing delegate would try to commission the same
device, which would fail.