MinMdns - do not reset hasIP/hasHostPort on partial failure
MinMdns can receive several packets and if any of the packets are ok,
a single valid ip or host/port is sufficient.

In practice, this was failing on mdns on ipv6 only builds when both ipv6
and ipv4 were sent in order. Specifically:
  - MDNS packet contains AAAA followed by A
  - Resolver parses AAAA and sets a valid IP
  - Resolver sees A and fails to parse (no IPv4 support) and resets
    valid IP

As a result, dns resolution was never calling complete (because it
thinks no valid IPs).