Store the updated fabric info when bringing up controller on existing fabric.
Right now bringing up a controller on an existing fabric updates the
in-memory fabric table but does not store the updated data.

This is not very consistent with our desire to not even keep the
fabric table in memory, and leads to inconsistent behavior if someone
later reloads the fabric table from storage in some way.

Specific changes:

1. Change FabricInfo::SetFabricInfo to not modify any member state until
   validation of the incoming info is complete, so we won't corrupt an existing
   fabric due to mistaken input and leave it in a bad state.

2. If updating the existing fabric in ProcessControllerNOCChain succeeds, store
   the updated fabric info.  This is done outside the fabric table for now,
   pending changes to the fabric table to store things itself instead of keeping
   them in RAM.